+++ product-service/src/test/java/n11bootcamp_project_backend/product_service/ProductServiceApplicationTests.java
@@ -1,0 +1,15 @@
+package n11bootcamp_project_backend.product_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+
+@Import(TestcontainersConfiguration.class)
+@SpringBootTest
+class ProductServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}