+++ product-service/src/test/java/n11bootcamp_project_backend/product_service/TestProductServiceApplication.java
@@ -1,0 +1,11 @@
+package n11bootcamp_project_backend.product_service;
+
+import org.springframework.boot.SpringApplication;
+
+public class TestProductServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.from(ProductServiceApplication::main).with(TestcontainersConfiguration.class).run(args);
+	}
+
+}